refactor(transaction-pool-service): rename package (#575)
* Rename package

* Empty commit

* Fix deeps check

* Fix p2p tests

* Fix e2e tests

* Fix workflow
