+++ packages/core/package.json
@@ -84,7 +84,7 @@
 		"@mainsail/proposer": "workspace:*",
 		"@mainsail/serializer": "workspace:*",
 		"@mainsail/state": "workspace:*",
-		"@mainsail/transaction-pool": "workspace:*",
+		"@mainsail/transaction-pool-service": "workspace:*",
 		"@mainsail/transactions": "workspace:*",
 		"@mainsail/utils": "workspace:*",
 		"@mainsail/validation": "workspace:*",
