+++ packages/transaction-pool-service/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@mainsail/transaction-pool",
+	"name": "@mainsail/transaction-pool-service",
 	"version": "0.0.1-alpha.17",
-	"description": "Transaction Pool Manager for Mainsail blockchain",
+	"description": "Transaction Pool Service for Mainsail blockchain",
 	"license": "GPL-3.0-only",
 	"contributors": [],
 	"type": "module",
