+++ packages/transaction-pool-service/source/storage.ts
@@ -7,7 +7,7 @@ import { ensureFileSync } from "fs-extra/esm";
 @injectable()
 export class Storage implements Contracts.TransactionPool.Storage {
 	@inject(Identifiers.ServiceProvider.Configuration)
-	@tagged("plugin", "transaction-pool")
+	@tagged("plugin", "transaction-pool-service")
 	private readonly configuration!: Providers.PluginConfiguration;
 
 	#database!: BetterSqlite3.Database;
